Add test for color detection
